requirements.txt
numpy==2.2.5
pandas==2.2.3
pyarrow==19.0.1
torch==2.6.0
tqdm==4.67.1
fastapi==0.115.0
uvicorn==0.32.0
scikit-learn==1.6.0
omegaconf==2.3.0
requests==2.32.5
python-dotenv==1.1.1
redis==5.0.1